[hotfix] make KeyByBenchmarks#IncreasingBaseSource more generic, creating BaseSourceWithKeyRange
